Move Lexer to lexer.h
